+++ stories-api/src/main/java/org/craftsmenlabs/stories/api/models/items/validated/ValidatedAcceptanceCriteria.java
@@ -15,11 +15,11 @@ import java.util.List;
 @NoArgsConstructor
 public class ValidatedAcceptanceCriteria extends AbstractValidatedItem<String> {
     @Builder
-    public ValidatedAcceptanceCriteria(List<Violation> violations, Rating rating, String item, float scoredPercentage, float missedPercentage, float scoredPoints, float missedPoints) {
+    public ValidatedAcceptanceCriteria(List<Violation> violations, Rating rating, String item, double scoredPercentage, double missedPercentage, double scoredPoints, double missedPoints) {
         super(violations, rating, item, scoredPercentage, missedPercentage, scoredPoints, missedPoints);
     }
 
     public static ValidatedAcceptanceCriteria empty(){
-        return new ValidatedAcceptanceCriteria(Collections.emptyList(), Rating.FAIL, "", 0f, 0f, 0f, 0f);
+        return new ValidatedAcceptanceCriteria(Collections.emptyList(), Rating.FAIL, "", 0, 0, 0, 0);
     }
 }
